refactored cleanup service

+++ src/ConversionCleanup.php
@@ -12,27 +12,13 @@ class ConversionCleanup
      * @var string
      */
     protected string $inputDisk;
-    
-    /**
-     * The output disk.
-     *
-     * @var string
-     */
-    protected string $outputDisk;
-    
-    /**
-     * The perform cleanup flag.
-     *
-     * @var bool
-     */
-    protected bool $performCleanup;
 
     /**
      * The cleanup strategy.
      *
      * @var string
      */
-    protected string $strategy;
+    protected bool $performCleanup;
 
     /**
      * Create a new ConversionCleanup instance.
@@ -43,61 +29,19 @@ class ConversionCleanup
     {
         $this->inputDisk = config('doxswap.input_disk');
 
-        $this->outputDisk = config('doxswap.output_disk');
-
-        $this->strategy = config('doxswap.cleanup_strategy'); // Cleanup strategies input, output, both, none
-    }
-
-    /**
-     * Cleanup the input file.
-     *
-     * @param string $inputFile
-     * @return void
-     */
-    protected function cleanupInput(string $inputFile): void
-    {
-        Storage::disk($this->inputDisk)->delete($inputFile);
-    }
-
-    /**
-     * Cleanup the output file.
-     *
-     * @param string $outputFile
-     * @return void
-     */
-    protected function cleanupOutput(string $outputFile): void
-    {
-        Storage::disk($this->outputDisk)->delete($outputFile);
-    }
-
-    /**
-     * Cleanup the input and output files.
-     *
-     * @param string $inputFile
-     * @param string $outputFile
-     * @return void
-     */
-    protected function cleanupInputAndOutput(string $inputFile, string $outputFile): void
-    {
-        $this->cleanupInput($inputFile);
-        
-        $this->cleanupOutput($outputFile);
+        $this->performCleanup = config('doxswap.perform_cleanup');
     }
 
     /**
      * Cleanup the input file and output files based on the strategy.
      *
      * @param string $inputFile
-     * @param string $outputFile
      * @return void
      */
-    public function cleanup(string $inputFile, string $outputFile): void
+    public function cleanup(string $inputFile): void
     {
-        match ($this->strategy) {
-            'input' => $this->cleanupInput($inputFile),
-            'output' => $this->cleanupOutput($outputFile),
-            'both' => $this->cleanupInputAndOutput($inputFile, $outputFile),
-            'none' => null,
-        };
+        if ($this->performCleanup) {
+            Storage::disk($this->inputDisk)->delete($inputFile);
+        }
     }
 }
